Add a 'utils.py' module containing all common, basic functions used through the Library.
